+++ 404.php
@@ -5,7 +5,7 @@
  * @link https://codex.wordpress.org/Creating_an_Error_404_Page
  *
  * @package Wordpress
- * @subpackage Coffee
+ * @subpackage Tea Coffee
  * @version 1.0.0
  */