+++ functions.php
@@ -5,14 +5,14 @@
  * @link https://developer.wordpress.org/themes/basics/theme-functions/
  *
  * @package Wordpress
- * @subpackage Coffee
+ * @subpackage Tea Coffee
  * @version 1.0.0
  */
 
 /**
  * Setup theme function
  */
-function coffee_setup() {
+function Tea Coffee_setup() {
 	// Add default posts and comments RSS feed links to head.
 	add_theme_support( 'automatic-feed-links' );
 
@@ -144,7 +144,7 @@ function WidgetLocation(){
         ));
 }
 
-add_action( 'after_setup_theme', 'coffee_setup' );
+add_action( 'after_setup_theme', 'Tea Coffee_setup' );
 add_action( 'wp_enqueue_scripts', 'wordpress_resources');
 add_action('customize_register', 'cfe_splash_info');
 add_filter( 'excerpt_length', 'custom_excerpt_length', 999 );