+++ home.php
@@ -5,7 +5,7 @@
  * @link https://codex.wordpress.org/Template_Hierarchy
  *
  * @package Wordpress
- * @subpackage Coffee
+ * @subpackage Tea Coffee
  * @version 1.0.0
  */