+++ index.php
@@ -21,10 +21,13 @@ get_header(); ?>
 if (have_posts()) :
 	while (have_posts()) : the_post();?>
 
-	<article class="post">
-		<h2><a href="<?php the_permalink() ?>"><?php the_title(); ?></a></h2>
-		<?php the_excerpt(); ?>
-	</article>
+	<section class="page-header">
+		<bubble>
+			<h1><?php the_title(); ?></h1>
+		</bubble>
+	</section>
+
+	<?php get_template_part( 'template-parts/page/content', 'page' ); ?>
 
 	<?php endwhile;
 	else :