+++ tea_coffee_core/footer.php
@@ -11,18 +11,18 @@
 ?>
 
 <footer class="main main-footer">
-        <container>
-            <nav class="footer-links">
-                <?php $args = array('theme_location' => 'footer'); ?>
-                <?php wp_nav_menu( $args ); ?>
-                <?php get_template_part( 'template-parts/site/logo', 'footer' ); ?>
-            </nav>
-        </container>
-    </footer>
-    <light></light>
-    <!--[if lt IE 9]>
-        <script src="js/scripts.js"></script>
-    <![endif]-->
-    <?php wp_footer(); ?>
+    <container>
+        <nav class="footer-links">
+            <?php $args = array('theme_location' => 'footer'); ?>
+            <?php wp_nav_menu( $args ); ?>
+            <?php get_template_part( 'template-parts/site/logo', 'footer' ); ?>
+        </nav>
+    </container>
+</footer>
+<light></light>
+<!--[if lt IE 9]>
+    <script src="js/scripts.js"></script>
+<![endif]-->
+<?php wp_footer(); ?>
 </body>
 </html>