+++ tea_coffee_core/template-parts/site/logo-header.php
@@ -10,10 +10,4 @@
  */
 ?>
 
-<?php 
-    if( has_custom_logo() ) { ?>
-        <icon class="header-icon">
-            <?php the_custom_logo(); ?>
-        </icon>
-    <?php }
-?>
+<?php the_custom_logo(); ?>