+++ tea_coffee_core/template-parts/site/site-head.php
@@ -16,10 +16,10 @@
     <meta charset="utf-8">
 
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="theme-color" content="#ffffff">
     <meta name="description" content="Homepage">
     <meta name="author" content="Technovus">
-    <link rel="shortcut icon" href="" type="image/x-icon" />
 
     <link href="https://fonts.googleapis.com/css?family=Audiowide|Kanit|Ubuntu|Raleway" rel="stylesheet">