+++ tea_coffee_core/home.php
@@ -11,22 +11,22 @@
 
 get_template_part( 'template-parts/site/site', 'head' ); ?>
 
-    <section class="news-header-section">
-		<? get_template_part( 'template-parts/header/analog', 'header'); ?>
+<section class="news-header-section">
+    <? get_template_part( 'template-parts/header/analog', 'header'); ?>
 
-        <section class="news-header-content">
-            <container>
-                <? get_template_part( 'template-parts/post/content', 'featured' ); ?>
-            </container>
-        </section>
-    </section>
-    <section class="news-page-content-section">
-        <container class="clearfix">
-            <posts>
-                <? get_template_part( 'template-parts/post/content', 'posts' ); ?>
-            </posts>
-            <? get_sidebar(); ?>
+    <section class="news-header-content">
+        <container>
+            <? get_template_part( 'template-parts/post/content', 'featured' ); ?>
         </container>
     </section>
+</section>
+<section class="news-page-content-section">
+    <container class="clearfix">
+        <posts>
+            <? get_template_part( 'template-parts/post/content', 'posts' ); ?>
+        </posts>
+        <? get_sidebar(); ?>
+    </container>
+</section>
 
 <? get_footer(); ?>