+++ tea_coffee_core/sidebar.php
@@ -6,7 +6,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.0.0
+ * @version 1.2.0
  */
 
 if ( ! is_active_sidebar( 'sidebar1' ) ) {