+++ tea_coffee_core/template-parts/post/content-posts.php
@@ -6,7 +6,7 @@
  * 
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.0.0
+ * @version 1.2.0
  */
 ?>
 
@@ -19,6 +19,23 @@ if (have_posts()) :
     <a href="<? the_permalink(); ?>">
         <container>
             <h2 class="post-title"><? the_title(); ?></h2>
+            <p class="post-info">
+                    <? echo the_time( 'F j, Y g:i a' ); ?> | Posted in 
+                    <?
+                        $categories = get_the_category();
+                        $seperator = ", ";
+                        $output = '';
+
+                        if ($categories) :
+                            foreach ($categories as $category) :
+
+                                $output .=  $category->cat_name . $seperator;
+
+                            endforeach;
+                            echo trim($output, $seperator);
+                        endif;
+                    ?>
+            </p>
             <div class="post-excerpt content">
                 <? the_excerpt(); ?>
             </div>