+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format of this change log is based on [Keep a Changelog](http://keepachangelog.com/)
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-
 ## [Unreleased] - XXXX-XXX-XX
+### Added
+- 
+
+## [1.2.0] - 2017-DEC-21
 ### Changed
 - Theme name to Tea Coffee
 - Index display format
@@ -38,4 +41,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - This CHANGELOG file
 - A CODE OF CONDUCT for community
 
-[Unreleased]: https://github.com/technovus-sfu/technovus-theme/compare/HEAD...HEAD
+
+## Compare
+[1.2.0]: https://github.com/technovus-sfu/technovus-theme/compare/v1.2.0...HEAD
+ 1.0.0 : https://github.com/technovus-sfu/technovus-theme/compare/v1.0.0...v1.2.0